Refactor profileUrl id extractor
