+++ src/features/broker-protection/extractors/profile-url.js
@@ -1,6 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { Extractor } from '../types.js'
-import { getIdFromProfileUrl } from '../actions/extract.js'
 
 /**
  * @implements {Extractor<{profileUrl: string; identifier: string} | null>}
@@ -22,7 +21,27 @@ export class ProfileUrlExtractor {
         }
 
         const profileUrl = strs[0]
-        profile.identifier = getIdFromProfileUrl(profileUrl, extractorParams.identifierType, extractorParams.identifier)
+        profile.identifier = this.getIdFromProfileUrl(profileUrl, extractorParams.identifierType, extractorParams.identifier)
         return profile
     }
+
+    /**
+     * Parse a profile id from a profile URL
+     * @param {string} profileUrl
+     * @param {import('../actions/extract.js').IdentifierType} identifierType
+     * @param {string} identifier
+     * @return {string}
+     */
+    getIdFromProfileUrl (profileUrl, identifierType, identifier) {
+        const parsedUrl = new URL(profileUrl)
+        const urlParams = parsedUrl.searchParams
+
+        // Attempt to parse out an id from the search parameters
+        if (identifierType === 'param' && urlParams.has(identifier)) {
+            const profileId = urlParams.get(identifier)
+            return profileId || profileUrl
+        }
+
+        return profileUrl
+    };
 }
